refactor: Drop SQLite

+++ internal/persisters/psql/communities.go
@@ -218,7 +218,7 @@ func (p *CommunitiesPersister) GetCommunities(
 	for _, community := range c {
 		cc = append(cc, persisters.Community{
 			ID:         community.ID,
-			Clients:    int64(community.Clients),
+			Clients:    community.Clients,
 			Persistent: community.Persistent,
 		})
 	}
@@ -249,7 +249,7 @@ func (p *CommunitiesPersister) CreatePersistentCommunity(
 
 	cc := &persisters.Community{
 		ID:         c.ID,
-		Clients:    int64(c.Clients),
+		Clients:    c.Clients,
 		Persistent: c.Persistent,
 	}
